Define domain to generate app-links from
Make Sharetribe read the app domain from ENV. The existing mechanism
doesn't suit our needs :see_no_evil:.

As explained in
https://github.com/coopdevs/sharetribe#setting-your-domain, this is what
Sharetribe uses to infer the URL to add in email links, for instance.
The relevant method is this:

```ruby
def full_domain(options= {})
  # assume that if port is used in domain config, it should
  # be added to the end of the full domain for links to work
  # This concerns usually mostly testing and development
  default_host, default_port = APP_CONFIG.domain.split(':')
  port_string = options[:port] || default_port

  if domain.present? && use_domain? # custom domain
    dom = domain
  else # just a subdomain specified
    dom = "#{self.ident}.#{default_host}"
    dom += ":#{port_string}" unless port_string.blank?
  end

  if options[:with_protocol]
    dom = "#{(APP_CONFIG.always_use_ssl.to_s == "true" ? "https://" : "http://")}#{dom}"
  end

  return dom
end
```

As you can see the method is not that straightforward. The problem is
that the only way to change the method's behaviour is by means of a data
migration and migrations are shared across environments. Therefore, when
running in production we would end up with a `community` record like
`community.domain = 'next.donalo.org'` leaving us without the option to
have sub-domain per environment.