Enable Stripe through a rake task and seed data
Sometimes we'll need to enable it for a server that already has a DB
setup and can't be dropped (like next.donalo.org right now). Often
times, we'll need to setup a new server from scratch (production or
staging) and the DB needs to be in this certain state: Stripe enabled.