Build the React bundle with webpack in development
I found out about the Procfiles and `client/` after successfully booting
the web container and seeing the error due to the missing
`vendor-bundle` and `app-bundle`. The comments there lead me to the
Procfiles where I managed to understand how assets are bundled in
development and came up with this solution.

All this was introduced in
https://github.com/sharetribe/sharetribe/pull/1918/ when Sharetribe
introduced React, and it's outlined in
https://github.com/coopdevs/sharetribe/tree/master/client.

There are a bunch of other Procfile services we are not covering yet but
this is a good start and gets the bundles in place.