Fix inconsistent marketplace seeding
I had to replace the manually created Community with the logic
Sharetribe uses to setup the marketplace from the UI. When you visit the
home page with an empty DB, the HomeController uses a couple services to
set things up. It does create more records than we did.

This, as a result, leaves you with a consistent development marketplace
but it does not recreate Donalo as I initially intended. It gives you an
onboarding progressed to the 14% percent (there's a top bar in the UI).
That can come after as a bunch of `update_attributes` calls on top of
this to store images and other customizations. Now I need this to
develop.

I also, needed to restore part of the logic we removed in
9356d70e9846be5c2cb5c0e1a8e7439e0dc35291.

Now, when you run `db:seed` you'll see the following message:

```
You can now navigate to your markeplace at http://0.0.0.0:3000?auth=_KOg0lTLSUc
```