Run Sphinx and DJ in the same container
Otherwise, `https://github.com/pat/ts-delayed-delta` can't find the
Sphinx binaries (indexer) to index whenever a listing is updated. They
stress it many times in the docs:

> Please keep in mind that you need to run the Delayed Job workers on
> the same machine as Sphinx, as direct access to Sphinx files (and the
> Sphinx command-line tools) is essential.

as well as in
https://freelancing-gods.com/thinking-sphinx/v4/deltas.html.

Note that we must therefore, detach `searchd` or the container run it
foreground and thus, never call `bundle exec rake jobs:work`. Docker has
some docs about that in
https://docs.docker.com/config/containers/multi-service_container/.