Mask `people` data in a DB view
This uses https://github.com/scenic-views/scenic. A well known gem from
Thoughtbot that adds tooling to manage DB view with regular migrations.

I also had to manually remove the line

```
/*!50013 DEFINER=`sharetribe`@`%` SQL SECURITY DEFINER */
```

from db/structure.sql so it works in CI. In there it fails with

```
Mysql2::Error: The user specified as a definer ('sharetribe'@'%') does not exist: SELECT "active_sessions" AS table_name,
```

because that DB user doesn't exist but it won't in production either.
According to
https://dev.mysql.com/doc/refman/5.7/en/stored-objects-security.html
though,

> If a definition omits the DEFINER attribute, the default definer is the user who creates the object.

And that seems fine as well.